xml config

+++ src/test/java/objects/OnSkipTestListener.java
@@ -2,8 +2,7 @@ package objects;
 
 import org.testng.*;
 
-import java.io.File;
-import java.io.IOException;
+import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.ArrayList;
@@ -12,13 +11,15 @@ import java.util.List;
 
 public class OnSkipTestListener implements IMethodInterceptor, ITestListener  {
 
-    private String path = System.getProperty("user.dir");
-    private File testFile;
+    private File testFile = new File(System.getProperty("user.dir") + "/src/test/resources/");
+    private String output = "";
+    private PrintWriter writeFile;
+
     @Override
     public List<IMethodInstance> intercept(List<IMethodInstance> list, ITestContext context) {
 
         String testName = context.getCurrentXmlTest().getName();
-        testFile = new File(path + "/src/test/resources/" + testName.substring(0, testName.lastIndexOf('_')) + "/" + testName + ".txt");
+        testFile = new File(testFile.getAbsolutePath() + "/" + testName.substring(0, testName.lastIndexOf('_')) + "/" + testName + ".txt");
 
         List<IMethodInstance> result = new ArrayList<>();
 
@@ -41,12 +42,35 @@ public class OnSkipTestListener implements IMethodInterceptor, ITestListener  {
     }
 
     @Override
-    public void onTestSuccess(ITestResult result) {
-        if (testFile.isFile()) {
+    public void onTestFailure(ITestResult result) {
+        output += result.getMethod().getMethodName() + "\n";
+    }
+
+    @Override
+    public void onFinish(ITestContext context) {
+        String s = context.getCurrentXmlTest().getName();
+        File testDir = new File(testFile.getParent());
+
+        File file = new File(testDir.getAbsolutePath() + "/" + context.getCurrentXmlTest().getName() + "_retest.txt");
+
+        if (output.length() > 0) {
+            if (!testDir.exists()) {
+                testDir.mkdirs();
+            }
             try {
-                Files.delete(Paths.get(testFile.getAbsolutePath()));
-            } catch (IOException e) {
+                writeFile = new PrintWriter(file, "UTF-8");
+            } catch (FileNotFoundException | UnsupportedEncodingException e) {
                 e.printStackTrace();
+            }
+            writeFile.print(output);
+            writeFile.close();
+        } else {
+            if (testFile.isFile()) {
+                try {
+                    Files.delete(Paths.get(testFile.getAbsolutePath()));
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
             }
         }
     }
